# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AgriPride Mkopo - Agricultural Loan Management Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /loans:
    get:
      operationId: listLoans
      tags: [loans]
      summary: List active loans
      parameters:
        - in: query
          name: status
          schema:
            type: string
          required: false
        - in: query
          name: farmerId
          schema:
            type: integer
          required: false
      responses:
        "200":
          description: List of loans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Loan"
  /loans/{id}:
    get:
      operationId: getLoan
      tags: [loans]
      summary: Get loan by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Loan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Loan"
        "404":
          description: Not found
  /loans/{id}/disburse:
    post:
      operationId: disburseLoan
      tags: [loans]
      summary: Disburse loan via M-Pesa B2C
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Disbursement initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DisburseResult"
  /loans/{id}/repayments:
    get:
      operationId: listRepayments
      tags: [repayments]
      summary: List repayments for a loan
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of repayments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Repayment"
  /loans/{id}/repay:
    post:
      operationId: initiateRepayment
      tags: [repayments]
      summary: Initiate M-Pesa STK push for loan repayment
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RepaymentInput"
      responses:
        "200":
          description: STK push sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StkPushResult"

  # ─── Dashboard ───────────────────────────────────────────────────────────────
...
  /mpesa/stk-callback:
    post:
      operationId: mpesaStkCallback
      tags: [mpesa]
      summary: M-Pesa STK push callback
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MpesaCallback"
      responses:
        "200":
          description: Acknowledged
  /mpesa/b2c-callback:
    post:
      operationId: mpesaB2cCallback
      tags: [mpesa]
      summary: M-Pesa B2C result callback
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MpesaCallback"
      responses:
        "200":
          description: Acknowledged
components:
  schemas:
...
    Loan:
      type: object
      required: [id, applicationId, farmerId, amount, interestRate, status, dueDate, createdAt]
      properties:
        id:            { type: integer }
        applicationId: { type: integer }
        farmerId:      { type: integer }
        farmerName:    { type: ["string", "null"] }
        farmerPhone:   { type: ["string", "null"] }
        amount:        { type: number }
        interestRate:  { type: number }
        status:        { type: string }
        dueDate:       { type: string }
        disbursedAt:   { type: ["string", "null"] }
        mpesaReceipt:  { type: ["string", "null"] }
        amountPaid:    { type: ["number", "null"] }
        createdAt:     { type: string }
    DisburseResult:
      type: object
      required: [message, conversationId]
      properties:
        message:        { type: string }
        conversationId: { type: string }

    # ─── Repayment ───────────────────────────────────────────────────────────
    Repayment:
      type: object
      required: [id, loanId, amount, paidAt]
      properties:
        id:           { type: integer }
        loanId:       { type: integer }
        amount:       { type: number }
        mpesaReceipt: { type: ["string", "null"] }
        paidAt:       { type: string }
    RepaymentInput:
      type: object
      required: [amount, phone]
      properties:
        amount: { type: number }
        phone:  { type: string }
    StkPushResult:
      type: object
      required: [message, checkoutRequestId]
      properties:
        message:           { type: string }
        checkoutRequestId: { type: string }

    # ─── Dashboard ───────────────────────────────────────────────────────────
...
    MpesaCallback:
      type: object
      properties:
        Body:
          type: object
          additionalProperties: true