# lib/api-zod/src/generated/types/stkPushResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AgriPride Mkopo - Agricultural Loan Management Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface StkPushResult {
  message: string;
  checkoutRequestId: string;
}